[STEP2] 로또(자동)(4) : MVC 로 분류
